LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/prefetch-single.https_from_protocol=http&to_protocol=http-expected.txt

CONSOLE MESSAGE: Prefetch request denied: URL must be secure (HTTPS)
CONSOLE MESSAGE: Prefetch request denied: URL must be secure (HTTPS)

PASS test single http url prefetch from a http url
